% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsamplr.R
\name{dsamplr}
\alias{dsamplr}
\title{Draw samples from a given distribution}
\usage{
dsamplr(n, distribution, ...)
}
\arguments{
\item{n}{A positive integer.}

\item{distribution}{A distribution ("normal", "poisson", "binomial")}

\item{...}{Additional parameters for the chosen distribution}
}
\value{
A dsamplr object
}
\description{
Draw samples from a given distribution
}
\examples{
dsamplr(100, "normal")
dsamplr(5000, "poisson", lambda=10)
}
