% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarydsamplr.R
\name{summary.dsamplr}
\alias{summary.dsamplr}
\title{Summary method for the S3 dsamplr object}
\usage{
\method{summary}{dsamplr}(object, ...)
}
\arguments{
\item{object}{A dsamplr object. Created by dsamplr function.}

\item{...}{Included for good practice in building generics.}
}
\value{
Summarises the number of samples and distribution.
}
\description{
Summary method for the S3 dsamplr object
}
